[Fix](func) Keep multi-match regex cache owner alive (#66909)
Problem Summary:

`multi_match_any` compiles each pattern array and stores the result in a
process-wide, direct-mapped cache.

The previous implementation retained the cache entry's shared_ptr only
while preparing the regex database and scratch space. The subsequent
scan used a **raw** `Regexps*`. If another concurrent request inserted a
different pattern array into the same cache bucket, the old cache entry
could be replaced and destroyed while it was still being scanned,
resulting in a use-after-free.

This change keeps the cache entry's shared_ptr alive for the entire
hs_scan operation. Cache eviction can still replace the global entry,
but the compiled database is not destroyed until all active scans have
completed.

Fix a potential use-after-free in multi_match_any when concurrent
pattern arrays collide in the compiled-regex cache.
