+++ be/src/exprs/function/regexps.h
@@ -67,24 +67,26 @@ private:
     ScratchPtr scratch;
 };
 
+using RegexpsPtr = std::shared_ptr<Regexps>;
+
 class DeferredConstructedRegexps {
 public:
     explicit DeferredConstructedRegexps(std::function<Regexps()> constructor_)
             : constructor(std::move(constructor_)) {}
 
-    Regexps* get() {
+    RegexpsPtr get() {
         std::lock_guard lock(mutex);
         if (regexps) {
-            return &*regexps;
+            return regexps;
         }
-        regexps = constructor();
-        return &*regexps;
+        regexps = std::make_shared<Regexps>(constructor());
+        return regexps;
     }
 
 private:
     std::mutex mutex;
     std::function<Regexps()> constructor;
-    std::optional<Regexps> regexps;
+    RegexpsPtr regexps;
 };
 
 using DeferredConstructedRegexpsPtr = std::shared_ptr<DeferredConstructedRegexps>;
